♻️ refactor: extract notification logic into separate helper module
